Add the ability to check for hook compatibility.
This will fail PRs that enable test_pull_requests enabled, and do not have enough permissions to setup the webhooks.
It will pass if ros-pull-request-builder has push access but not admin access to view the webhooks.
In this case it is unknown if the user setup the webhooks manually or not.